% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{calculate_density}
\alias{calculate_density}
\title{Estimate weighted kernel density}
\usage{
calculate_density(w, x, method, adjust = 1, map = TRUE)
}
\arguments{
\item{w}{Vector with weights for each observation}

\item{x}{Matrix with dimensions where to calculate the density from. Only
the first two dimensions will be used}

\item{method}{Kernel density estimation method:
\itemize{
\item \code{ks}: Computes density using the \code{kda} function from the
 \code{ks} package.
Ideal for small-medium datasets (up to 70,000 cells in a desktop computer).
 Use other methods for larger datasets.
\item \code{wkde}: Computes density using a modified version of the
 \code{kde2d} function from the \code{MASS}
package to allow weights. Bandwidth selection from the \code{ks} package
 is used instead.
\item \code{sm}: Computes density using the \code{sm.density} function from
 the \code{sm} package
}}

\item{adjust}{Numeric value to adjust to bandwidth. Default: 1. Not available
for \code{ks} method}

\item{map}{Whether to map densities to individual observations}
}
\value{
If \code{map} is \code{TRUE}, a vector with corresponding densities
 for each observation is returned. Otherwise,
a war a list with the density estimates from the selected method is returned.
}
\description{
Estimate weighted kernel density
}
\examples{

dens <- Nebulosa:::calculate_density(iris[, 3], iris[, 1:2], method = "wkde")
}
\author{
Jose Alquicira-Hernandez
}
