% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_density_}
\alias{plot_density_}
\title{Plot density estimates}
\usage{
plot_density_(
  z,
  feature,
  cell_embeddings,
  dim_names,
  shape,
  size,
  legend_title,
  pal = c("viridis", "magma", "cividis", "inferno", "plasma")
)
}
\arguments{
\item{z}{Vector with density values for each cells}

\item{feature}{Name of the feature being plotted}

\item{cell_embeddings}{Matrix with cell embeddings}

\item{dim_names}{Names of the dimensions from the cell embeddings}

\item{shape}{Geom shape}

\item{size}{Geom size}

\item{legend_title}{String used as legend title}

\item{pal}{String specifying the viridis color palette to use}
}
\value{
A ggplot object
}
\description{
Plot density estimates
}
\author{
Jose Alquicira-Hernandez
}
